gleam-skin-main/backend/requirements.txt
scikit-learn
joblib
pandas
django
djangorestframework
django-cors-headers
huggingface_hub
python-dotenv
psycopg2-binary
bcrypt
certifi
pytest
pytest-django
pytest-cov
gunicorn
dj-database-url
whitenoise
